Fix runtime in mind.dm

+++ code/datums/mind.dm
@@ -93,8 +93,9 @@
 	if(new_character.mind)		//remove any mind currently in our new body's mind variable
 		new_character.mind.current = null
 
-	if(current.client)
-		current.client.destroy_UI()
+	if (current) // check if current exists otherwise we cannot read current.client
+		if(current.client)
+			current.client.destroy_UI()
 
 	current = new_character		//link ourself to our new body
 	new_character.mind = src	//and link our new body to ourself
